+++ vision/transforms/center_crop_test.go
@@ -6,6 +6,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"gocv.io/x/gocv"
 )
 
 func TestCenterCrop(t *testing.T) {
@@ -13,23 +14,28 @@ func TestCenterCrop(t *testing.T) {
 
 	blue := color.RGBA{0, 0, 255, 255}
 	m := drawImage(image.Rect(0, 0, 2, 2), blue)
+	mCv, _ := gocv.ImageToMatRGB(m)
 
-	r := CenterCrop(2, 2).Run(m)
+	rCv := CenterCrop(2, 2).Run(mCv)
+	r, _ := rCv.ToImage()
+
 	a.Equal(image.Rectangle{Min: image.Point{0, 0}, Max: image.Point{2, 2}}, r.Bounds())
 	a.True(colorEqual(blue, r.At(0, 0)))
 	a.True(colorEqual(blue, r.At(0, 1)))
 	a.True(colorEqual(blue, r.At(1, 0)))
 	a.True(colorEqual(blue, r.At(1, 1)))
 
-	r = CenterCrop(1, 2).Run(m)
+	rCv = CenterCrop(1, 2).Run(mCv)
+	r, _ = rCv.ToImage()
 	a.Equal(image.Rectangle{Min: image.Point{0, 0}, Max: image.Point{2, 1}}, r.Bounds())
 	a.True(colorEqual(blue, r.At(0, 0)))
 	a.True(colorEqual(blue, r.At(1, 0)))
 
-	r = CenterCrop(1, 1).Run(m)
+	rCv = CenterCrop(1, 1).Run(mCv)
+	r, _ = rCv.ToImage()
 	a.Equal(image.Rectangle{Min: image.Point{0, 0}, Max: image.Point{1, 1}}, r.Bounds())
 	a.True(colorEqual(blue, r.At(0, 0)))
 
-	a.Panics(func() { CenterCrop(3, 1).Run(m) })
-	a.Panics(func() { CenterCrop(1, 3).Run(m) })
+	a.Panics(func() { CenterCrop(3, 1).Run(mCv) })
+	a.Panics(func() { CenterCrop(1, 3).Run(mCv) })
 }
